requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-jazzmin==2.6.0
django-oauth-toolkit==2.3.0
